requirements.txt
PyQt5==5.15.9
PyPDF2==3.0.1
markdown==3.4.4
Pillow==10.0.0
configparser==5.3.0